+++ src/JavLuv/MainWindowViewModel.cs
@@ -63,7 +63,8 @@ namespace JavLuv
             }
 
             var mainView = Application.Current.MainWindow as MainWindow;
-            m_movieCollection = new MovieCollection(Application.Current.Dispatcher, IsReadOnlyMode);
+            m_movieCollection = new MovieCollection(
+                Application.Current.Dispatcher, IsReadOnlyMode, JavLuv.Settings.Get().UseJapaneseNameOrder);
             m_movieScanner = new MovieScanner(m_movieCollection, mainView.webView);
             m_settingsViewModel = new SettingsViewModel(this);
             m_reportViewModel = new ReportViewModel(this);
@@ -349,8 +350,13 @@ namespace JavLuv
             }
 
             // Optionally move/rename post-scan
-            if (JavLuv.Settings.Get().EnableMoveRename && JavLuv.Settings.Get().MoveRenameAfterScan && m_movieScanner.IsCancelled == false)
+            if (m_movieScanner.IsRescanFiles == false && 
+                JavLuv.Settings.Get().EnableMoveRename && 
+                JavLuv.Settings.Get().MoveRenameAfterScan && 
+                m_movieScanner.IsCancelled == false)
+            {
                 m_movieBrowserViewModel.MoveRenameMovies(m_movieScanner.Movies);
+            }
 
             // Add new movies to movie database
             if (JavLuv.Settings.Get().AddToCollection)
@@ -488,6 +494,7 @@ namespace JavLuv
 
         public void StartRecan(List<string> scanDirectories)
         {
+            m_movieScanner.IsRescanFiles = true;
             m_movieScanner.Start(scanDirectories);
             SidePanel.SettingsIsEnabled = false;
             NotifyPropertyChanged("IsScanning");